refactor queueAdmin into abstract class.
